Allow usage of different schemas for documents
This commit adds support for using dedicated schemas for storing
documents. This way one can separate storage of documents of different
scopes do dedicated schemas per scope. THis allows for a better
structuring of the documents within the database.

This commit does **not** handle creating the schemas. The schema needs
to be available to the DocumentStore for this to work.

Basically this allows one to set the schema right within the table-name
by separating schema and tablename by a dot like this:
`schemaname.tablename`. The name will then be split on the dot and the
former part be used as schema-name, the later part as table-name. Any
table-prefixes will only be added to the table name, not to the scheme.