Fix issue with empty array for IN-clause
When an IN-clause is created with an empty array PostgreSQL will blow up
with an SQL-Error as the resulting SQL would be ` x IN()` which is not
allowed.

This commit works around that by returning a query-part that always
returns
false as querying for something that is not part of something else will
always return false.